* commit 'b9fa7a8cbca5e438cb0385fe8f87f1f2c63ad322': (27 commits)
  add的文件
  权限存取
  调试日志
  居中布局性能优化
  bug
  bug
  bug
  bug
  bug
  bug
  BI BUG
  指标弹出明细表
  ETL字段信息
  bug
  过滤取字段
  group添加position属性test
  树控件多对多逻辑
  调整列宽
  bug
  bug
  ...

+++ fbi_engine_base/src/com/fr/bi/cal/stable/cube/file/DateColumnFile.java
@@ -101,7 +101,7 @@ public class DateColumnFile implements ColumnFile<Long> {
     }
 
     @Override
-    public void addDataValue(long row, Long value) {
+    public void addDataValue(int row, Long value) {
         base.addDataValue(row, value);
     }
